Extracted extension to comply with max file length rules
